utils.py renamed to core_utils.py
