+++ container_service_extension/client/cse.py
@@ -13,7 +13,7 @@ from container_service_extension.client.ovdc_commands import ovdc_group
 from container_service_extension.client.system_commands import system_group
 from container_service_extension.client.template_commands import template_group
 from container_service_extension.logging.logger import CLIENT_LOGGER
-import container_service_extension.common.utils.utils as utils
+import container_service_extension.common.utils.core_utils as utils
 
 
 @vcd.group(short_help='Manage Kubernetes clusters',
