+++ container_service_extension/client/de_cluster_tkg.py
@@ -23,7 +23,7 @@ import container_service_extension.exception.exceptions as cse_exceptions
 import container_service_extension.logging.logger as logger
 import container_service_extension.common.utils.pyvcloud_utils as vcd_utils
 import container_service_extension.common.constants.shared_constants as shared_constants
-import container_service_extension.common.utils.utils as utils
+import container_service_extension.common.utils.core_utils as utils
 
 
 class DEClusterTKG:
