+++ container_service_extension/common/utils/core_utils.py
