+++ container_service_extension/common/utils/vsphere_utils.py
@@ -13,7 +13,7 @@ from pyvcloud.vcd.vm import VM
 from vsphere_guest_run.vsphere import VSphere
 
 from container_service_extension.logging.logger import NULL_LOGGER
-from container_service_extension.common.utils.utils import NullPrinter
+from container_service_extension.common.utils.core_utils import NullPrinter
 
 cache = LRUCache(maxsize=1024)
 vsphere_list = []
