+++ container_service_extension/installer/right_bundle_manager.py
@@ -10,7 +10,7 @@ from container_service_extension.logging.logger import NULL_LOGGER
 from container_service_extension.logging.logger import SERVER_CLOUDAPI_WIRE_LOGGER
 import container_service_extension.common.utils.pyvcloud_utils as vcd_utils
 from container_service_extension.common.constants.shared_constants import RequestMethod
-import container_service_extension.common.utils.utils as utils
+import container_service_extension.common.utils.core_utils as utils
 
 
 class RightBundleManager():
