+++ container_service_extension/installer/templates/remote_template_manager.py
@@ -11,8 +11,8 @@ import yaml
 import container_service_extension.installer.templates.local_template_manager as ltm
 from container_service_extension.logging.logger import NULL_LOGGER
 from container_service_extension.common.constants.server_constants import ScriptFile
-from container_service_extension.common.utils.utils import download_file
-from container_service_extension.common.utils.utils import NullPrinter
+from container_service_extension.common.utils.core_utils import download_file
+from container_service_extension.common.utils.core_utils import NullPrinter
 
 
 REMOTE_TEMPLATE_COOKBOOK_FILENAME = 'template.yaml'
