+++ container_service_extension/rde/acl_service.py
@@ -15,7 +15,7 @@ import container_service_extension.rde.models as def_models
 import container_service_extension.common.utils.pyvcloud_utils as vcd_utils
 import container_service_extension.common.constants.server_constants as server_constants
 import container_service_extension.common.constants.shared_constants as shared_constants
-import container_service_extension.common.utils.utils as utils
+import container_service_extension.common.utils.core_utils as utils
 
 
 class ClusterACLService:
