+++ container_service_extension/rde/cluster_service.py
@@ -49,7 +49,7 @@ import container_service_extension.lib.telemetry.constants as telemetry_constant
 import container_service_extension.lib.telemetry.telemetry_handler as telemetry_handler  # noqa: E501
 import container_service_extension.common.thread_local_data as thread_local_data
 import container_service_extension.common.utils.thread_utils as thread_utils
-import container_service_extension.common.utils.utils as utils
+import container_service_extension.common.utils.core_utils as utils
 import container_service_extension.common.utils.vsphere_utils as vs_utils
