+++ container_service_extension/rde/ovdc_service.py
@@ -28,7 +28,7 @@ from container_service_extension.lib.telemetry.constants import PayloadKey
 import container_service_extension.lib.telemetry.telemetry_handler as telemetry_handler # noqa: E501
 import container_service_extension.common.thread_local_data as thread_local_data
 import container_service_extension.common.utils.thread_utils as thread_utils
-import container_service_extension.common.utils.utils as utils
+import container_service_extension.common.utils.core_utils as utils
 
 
 def update_ovdc(operation_context: ctx.OperationContext,
