+++ container_service_extension/server/request_handlers/pks_ovdc_handler.py
@@ -32,7 +32,7 @@ from container_service_extension.lib.telemetry.telemetry_handler import record_u
 from container_service_extension.lib.telemetry.telemetry_handler import record_user_action_details  # noqa: E501
 from container_service_extension.lib.telemetry.telemetry_handler import record_user_action_telemetry  # noqa: E501
 import container_service_extension.common.thread_local_data as thread_local_data
-import container_service_extension.common.utils.utils as utils
+import container_service_extension.common.utils.core_utils as utils
 
 SYSTEM_DEFAULT_COMPUTE_POLICY_NAME = "System Default"
